requirements.txt
gym==0.18.0
pytorch==1.8.1